requirements.txt
azure.cognitiveservices.speech
requests
pydub